ensure widgets can be garbage collected
To be more test friendly.

Widgets (qargparse widgets in this case) were defined
as class attributes (not in __init__), so it won't be
garbage collected after main window closed, and signals
are still connected. And somehow they still able to
trigger the model which is created in next test session
to do model reset. So how many tests has ran, how many
reset signal will be emitted, to the same model, hence
the test can easily got segfault (especially Linux).